fix : 빌드 에러, generatemetadata는 use client에서 사용 못함
